requirements.txt
pymongo
flask
python-dotenv

membre du groupe: 